fix(ci): resolve phpstan + swift-format regressions
PHPStan (4 errors):
- EntitlementsController: ?->plan_key unreachable under auth:sanctum;
  use -> and rely on the guard.
- ReminderController: cast now()->valueOf() to int — Carbon returns float
  on some platforms.
- Plan / Reminder: add @use HasFactory<*Factory> generic annotation;
  introduce PlanFactory so the annotation has a target.

swift-format (CI runs older parser than local Xcode):
- Strip trailing commas in function-call argument lists across the new
  files. Swift 6 supports them but macos-14 swift-format parses them as
  syntax errors.

+++ api/app/Http/Controllers/Api/V1/ReminderController.php
@@ -50,7 +50,7 @@ class ReminderController extends Controller
         $reminder = $request->user()->reminders()->create([
             'time_local' => $data['time_local'],
             'enabled' => $data['enabled'] ?? true,
-            'updated_at_ms' => now()->valueOf(),
+            'updated_at_ms' => (int) now()->valueOf(),
         ]);
 
         return response()->json($reminder, 201);
@@ -66,7 +66,7 @@ class ReminderController extends Controller
         ]);
 
         $reminder->fill($data);
-        $reminder->updated_at_ms = now()->valueOf();
+        $reminder->updated_at_ms = (int) now()->valueOf();
         $reminder->save();
 
         return response()->json($reminder);
